changes following review
