+++ tests/next_tests/integration_tests/feature_tests/cases.py
@@ -29,7 +29,7 @@ from gt4py.eve import extended_typing as xtyping
 from gt4py.eve.extended_typing import Self
 from gt4py.next import common
 from gt4py.next.ffront import decorator
-from gt4py.next.iterator import embedded, ir as itir
+from gt4py.next.iterator import embedded
 from gt4py.next.program_processors import processor_interface as ppi
 from gt4py.next.type_system import type_specifications as ts, type_translation
 
@@ -424,15 +424,6 @@ def verify_with_default_data(
 
 
 @pytest.fixture
-def no_default_backend():
-    """Temporarily switch off default backend for feature tests."""
-    backup_backend = decorator.DEFAULT_BACKEND
-    decorator.DEFAULT_BACKEND = no_backend
-    yield
-    decorator.DEFAULT_BACKEND = backup_backend
-
-
-@pytest.fixture
 def cartesian_case(no_default_backend, fieldview_backend):  # noqa: F811 # fixtures
     yield Case(
         fieldview_backend,
@@ -548,11 +539,6 @@ def get_default_data(
         param_types.pop("out")
     inps = tuple(allocate(case, fieldview_prog, name)() for name in param_types)
     return inps, kwfields
-
-
-def no_backend(program: itir.FencilDefinition, *args: Any, **kwargs: Any) -> None:
-    """Temporary default backend to not accidentally test the wrong backend."""
-    raise ValueError("No backend selected! Backend selection is mandatory in tests.")
 
 
 @dataclasses.dataclass
